Finish and test the CustomerController just need to add the hashing method and the password attribute to the User class

+++ Application/Services/CustomerService.cs
@@ -20,38 +20,42 @@ public class CustomerService : ICustomerService
 
     public Customer AddCustomer(User user, Customer customer)
     {
-        var returnedUser = userRepository.GetFirstOrDefault(x => x.UserName == user.UserName 
-        && x.PasswordHash == user.PasswordHash);
+        //var returnedUser = userRepository.GetFirstOrDefault(x => x.UserName == user.UserName 
+        //&& x.PasswordHash == user.PasswordHash);
 
-        if (returnedUser == null)
-        {
-            returnedUser = userRepository.Add(user);
-        }
+        //if (returnedUser == null)
+        //{
+        //    returnedUser = userRepository.Add(user);
+        //}
 
-        customer.UserId = returnedUser.Id;
+        //customer.UserId = returnedUser.Id;
 
         var returnedCustomer = customerRepository.Add(customer);
 
         customerRepository.SaveChanges();
+
+        //returnedCustomer.User = returnedUser;
 
         return returnedCustomer;
     }
 
     public async Task<Customer> AddCustomerAsync(User user, Customer customer)
     {
-        var returnedUser = await userRepository.GetFirstOrDefaultAsync(x => x.UserName == user.UserName
-        && x.PasswordHash == user.PasswordHash);
+        //var returnedUser = await userRepository.GetFirstOrDefaultAsync(x => x.UserName == user.UserName
+        //&& x.PasswordHash == user.PasswordHash);
 
-        if (returnedUser == null)
-        {
-            returnedUser = await userRepository.AddAsync(user);
-        }
+        //if (returnedUser == null)
+        //{
+        //    returnedUser = await userRepository.AddAsync(user);
+        //}
 
-        customer.UserId = returnedUser.Id;
+        //customer.UserId = returnedUser.Id;
 
         var returnedCustomer = await customerRepository.AddAsync(customer);
 
         await customerRepository.SaveChangesAsync();
+
+        //returnedCustomer.User = returnedUser;
 
         return returnedCustomer;
     }
@@ -121,8 +125,16 @@ public class CustomerService : ICustomerService
         var customer = customerRepository.Remove(id);
         customerRepository.SaveChanges();
 
-        userRepository.Remove(id);
+        if (customer == null)
+            return null;
+
+        var user = userRepository.Remove(id);
         userRepository.SaveChanges();
+
+        if (user == null)
+            return null;
+
+        customer.User = user;
 
         return customer;
     }
@@ -132,15 +144,73 @@ public class CustomerService : ICustomerService
         var customer = customerRepository.Remove(id);
         await customerRepository.SaveChangesAsync();
 
-        userRepository.Remove(id);
+        if (customer == null)
+            return null;
+
+        var user = userRepository.Remove(id);
         await userRepository.SaveChangesAsync();
 
+        if(user == null)
+            return null;
+
+        customer.User = user;
+
         return customer;
+    }
+
+    public IEnumerable<Customer>? GetAllCustomers()
+    {
+        var customers = customerRepository.GetAll();
+
+        if(customers == null)
+            return null;
+
+        foreach (var customer in customers)
+        {
+            var user = userRepository.GetById(customer.UserId);
+
+            if(user == null)
+                continue;
+
+            customer.User = user;
+        }
+
+        return customers;
+    }
+
+    public async Task<IEnumerable<Customer>?> GetAllCustomersAsync()
+    {
+        var customers = await customerRepository.GetAllAsync();
+
+        if (customers == null)
+            return null;
+
+        foreach (var customer in customers)
+        {
+            var user = await userRepository.GetByIdAsync(customer.UserId);
+
+            if (user == null)
+                continue;
+
+            customer.User = user;
+        }
+
+        return customers;
     }
 
     public Customer? GetCustomerById(Guid id)
     {
         var customer = customerRepository.GetById(id);
+
+        if (customer == null) 
+            return null;
+
+        var user = userRepository.GetById(id);
+
+        if (user == null) 
+            return null;
+
+        customer.User = user;
 
         return customer;
     }
@@ -148,6 +218,16 @@ public class CustomerService : ICustomerService
     public async Task<Customer?> GetCustomerByIdAsync(Guid id)
     {
         var customer = await customerRepository.GetByIdAsync(id);
+
+        if (customer == null)
+            return null;
+
+        var user = await userRepository.GetByIdAsync(id);
+
+        if (user == null)
+            return null;
+
+        customer.User = user;
 
         return customer;
     }
@@ -193,6 +273,16 @@ public class CustomerService : ICustomerService
         var returnedCustomer = customerRepository.Update(id, customer);
         customerRepository.SaveChanges();
 
+        if (returnedCustomer == null)
+            return null;
+
+        var user = userRepository.GetById(id);
+
+        if (user is null) 
+            return null;
+
+        returnedCustomer.User = user;
+
         return returnedCustomer;
     }
 
@@ -200,6 +290,16 @@ public class CustomerService : ICustomerService
     {
         var returnedCustomer = await customerRepository.UpdateAsync(id, customer);
         await customerRepository.SaveChangesAsync();
+
+        if (returnedCustomer == null)
+            return null;
+
+        var user = await userRepository.GetByIdAsync(id);
+
+        if (user is null)
+            return null;
+
+        returnedCustomer.User = user;
 
         return returnedCustomer;
     }
